Stop hiding System.Collections.Immutable.dll from the compiler
`Compile` has been held private since the PackageReference was first added, and no clue was left as to why this was a good idea. Nowadays though it's almost certainly a bad idea, because we probably do (or will) declare public APIs that referenc types from this dependency, and not exposing the dependency to the compiler will produce compile errors for our users.