IRemoteServiceBroker should communicate the full pipe name
The issue was that only the leaf name of the pipe was communicated over the protocol. On Windows, this was fine because the leaf name is the full name. But on linux/mac, the pipe is actually a file on disk, so the full path is required. Prior to this change, only the leaf name was communicated anyway, and both processes had to agree (out of band) on what the prefixed path should be. This was fundamentally broken as the protocol should handle this. It was a keen problem because there are arguably any number of paths that might be appropriate to use as a prefix.

This fixes the last of the node test failures on linux.