Accommodate Windows pipe names w/o prefix
The NPM package required the `\\.\pipe\` prefix to be returned by an `IRemoteServiceBroker` to connect to a pipe on Windows. Most `IRemoteServiceBroker` implementations in use today do not include this prefix. While the prefix is arguably a good thing to include, we should accommodate them when they don't include it.