Omit the pipe name prefix for remote services on Windows
Although its arguably a good thing to fully-qualify all pipe names, older Microsoft.ServiceHub.Framework builds assumed the prefix on Windows would not be included in the `IRemoteServiceBroker`-returned pipe name and fail when it is included.
To help us migrate to a fully-qualified world, the new code can *consume* Windows pipe names with or without the prefix. And for the short term this change rolls back the behavior of what is returned to once again omit the prefix (only on Windows) so that older clients won't fail.

When all the clients have upgraded, we should revert this particular commit and return fully-qualified pipe names all the time.