Fix args in notifications raised by javascript
A single argument to a notification was being serialized as the entire JSON-RPC `params` value, which per the protocol meant instead of one positional argument, we were sending the argument's properties each as their own named argument.
With this fix, we more carefully work with the vscode-jsonrpc library to ensure the arguments are always retained by the protocol as they were intended.

Use case:

```ts
const args = { seeds: 5, weight: 8 }
this.emit('picked', args)
```

This *was* being transmitted as:

```json
{
	"method": "picked",
	"params": {
		"seeds": 5,
		"weight": 8
	}
}
```

Note how the `args` object itself is directly used as the value for `params`, so this is interpreted by the protocol as two named arguments.
With the fix, it is now being transmitted as:

```json
{
	"method": "picked",
	"params": [
		{
			"seeds": 5,
			"weight": 8
		}
	]
}
```

Note the surrounding array, which means we're using *positional* arguments.
There is only one argument, which has two properties.
This now accurately represents what was actually done in the event raising code and means that a receiver can now dispatch to a method that was expecting the same signature (one arg, two properties).