Fix brokered service query failures when multiple requests come in early
We had a bug where if a brokered service is registered with a proferring callback, only the *first* request would invoke and wait on that callback. All subsequent requests would skip over that and fail if the callback hadn't finished by proffering the service factory yet.
The fix is to ensure that while only the first request should *invoke* the callback, *all* requests should await the result of that callback (if in fact it is awaitable) so that no request fails simply because it was made before the service was proferred.