Update OSS Packages for Compliance 2024
Update OSS Packages for Compliance 2024.

The latest versions of tslint and eslint are both older than 4 years, which would categorize them as "Abandoned". In order to stay compliant we would need to take ownership of those packages/repos. To avoid having to that they were removed from the package.json.

All other updates are moving any dependencies that are older than 18 months to the latest stable versions.