Rename `AdditionalClientInterfaces` to `AdditionalServiceInterface`
Because after all, the collection describes interfaces that the *service* is expected to implement.
While the purpose of the property is to inform which interfaces the *client proxy* should implement, in the event that we offer the reverse as well where client callbacks may have multiple interfaces, it will be important to be able to distinguish which interfaces are on the client vs. the service. And I'm thinking this wording will be more clear.