Add interfaces to the client proxy based on brokered service registration
When a brokered service is registered, it now has the opportunity to identify which optional interfaces it implements so that its client will be able to simply cast the proxy to that interface to use it (or test for it). This registration uses type _names_ rather that Type objects themselves so as to not defeat lazy loading of the assemblies that may declare these interfaces.

When used in Visual Studio, this change will need to be combined with a change to the VS host to offer this registration data and read it at startup.

Closes #225