Add support for MEF exported brokered services with null versions
Null versioned services are supposed to serve as a "catch-all" for any version the client might request. This worked for VSPackage-proffered services, but MEF exported services failed to be activated with arbitrary versions. This change corrects that.

This change also enables exported brokered services to "reject" requests by returning a null descriptor, which is important with catch-all versioning so that unknown versions can be rejected at runtime.