Add support for the new StreamJsonRpc formatter that supports Nerdbank.MessagePack (#369)
* Add support for the new StreamJsonRpc formatter that supports Nerdbank.MessagePack

This is quite an intense undertaking, because Nerdbank.MessagePack introduces our first NativeAOT-safe serializer, which brings in new requirements and usage restrictions. But it is also a great opportunity. Here's all what's in this PR:

* Source generated local proxies. Dynamic proxies still exist, but the new PolyType descriptor only supports activating source generated proxies. Source generated proxies will start faster (no Ref.Emit or JIT) and are NativeAOT safe.
* Better runtime type checking for what optional interfaces a brokered service implements. We share a common interface with StreamJsonRpc now (which also recently added source generated proxy support) so clients can perform service interface testing in one consistent way whether the proxy is a remote or local one.
* This library now advertises NativeAOT safety. A great deal of APIs are in fact *not* NativeAOT safe though and these now are annotated as requiring dynamic or unreferenced code. At the moment, so many APIs are annotated that ultimately nothing in the library is useful in a NativeAOT environment. This is a point-in-time state. Once we get a NativeAOT-safe JSON formatter for our Framework service protocols to use, most of these attributes will be removed, unblocking use in a NativeAOT app.

This change required a whole new descriptor type that does *not* derive from `ServiceJsonRpcDescriptor` because `ITypeShapeProvider` is a required property on a descriptor now and because the pre-existing descriptor class offered functionality that is not NativeAOT-safe. But the new class derives from the same abstract base class as the old one did, so it's still compatible with all other APIs.

* Touch-up in response to copilot feedback

* Accept a binary breaking change of converting a virtual method to abstract

This is acceptable because we don't expect anyone to be implementing their own descriptors outside this assembly, and we need to move dynamic runtime behavior to a derived type.

* Fix events, and add more tests

* Update LKG test output

* Fix more tests

* Fix support for multiple service interfaces

* Fix line ending test failures on linux

* Skip line ending format checks which fail on linux

* CR feedback

* Add delegating variant of the polytype descriptor