Fix `IMissingServiceDiagnosticsService` descriptor and `ServiceMoniker` serialization
The descriptor for `IMissingServiceDiagnosticsService` was evidently always broken, because is used MessagePack-CSharp and `MissingServiceAnalysis` was not annotated with `[MessagePackObject]`.
The fix then is to just switch to Nerdbank.MessagePack, which doesn't require the attributes.

Also `ServiceMoniker` was dropping its `Version` property when serialized with Nerdbank.MessagePack on account of PolyType not seeing the constructor that could initialize the property.