Redesign how the new polytype descriptor works
The initial design lacked support for client callback interfaces. Since PolyType-based serialization requires shapes for both client and service interfaces, and since we combine sets of interfaces, we need to distinguish the types/shapes for service vs. client.
So one big theme change is that we now have different methods for creating proxies or adding RPC targets for each of the two roles.
The other theme change is that we have a _set_ of definite interfaces for services and clients, and another set of _optional_ interfaces for the service.
While the optional set is ignored, it stores the shapes that may be added to the set of _active_ interfaces based on whatever policy the container wants. In the case of VS, it's brokered service registration that names the types to be added, which can then be used to find matches among the optional shapes.